refactor: import context window helpers from modelConfig directly
