+++ frontend/src/lib/components/copilot/autocomplete/Autocompletor.ts
@@ -3,7 +3,8 @@ import { sleep } from '$lib/utils'
 import { editor as meditor, Position, languages, type IDisposable } from 'monaco-editor'
 import { LRUCache } from 'lru-cache'
 import { autocompleteRequest } from './request'
-import { FIM_MAX_TOKENS, getModelContextWindow } from '../lib'
+import { FIM_MAX_TOKENS } from '../lib'
+import { getModelContextWindow } from '../modelConfig'
 import { setGlobalCSS } from '../shared'
 import { supportsAutocomplete } from '../utils'
 import { get } from 'svelte/store'
