+++ frontend/src/lib/components/copilot/chat/AIChatManager.svelte.ts
@@ -36,7 +36,8 @@ import { loadApiTools } from './api/apiTools'
 import { prepareScriptUserMessage } from './script/core'
 import { prepareNavigatorUserMessage } from './navigator/core'
 import { sendUserToast } from '$lib/toast'
-import { getKnownModelContextWindow, workspaceAIClients } from '../lib'
+import { workspaceAIClients } from '../lib'
+import { getKnownModelContextWindow } from '../modelConfig'
 import { dfs } from '$lib/components/flows/previousResults'
 import { getStringError } from './utils'
 import { type PasteAttachment } from './pasteTokens'
