+++ frontend/src/lib/components/copilot/chat/AIChatManager.test.ts
@@ -50,8 +50,6 @@ vi.mock('$lib/aiStore', () => ({
 }))
 
 vi.mock('../lib', () => ({
-	getModelContextWindow: () => 128000,
-	getKnownModelContextWindow: () => 128000,
 	workspaceAIClients: {
 		subscribe: () => () => undefined,
 		getOpenaiClient: mocks.getOpenaiClient,
